Added chosen list to ListInput
